Refactor entity editor to new file
